services/text_Analysis/requirements.txt
<<<<<<< HEAD
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
=======
fastapi==0.104.1
uvicorn==0.24.0
>>>>>>> Text-Analysis
pymongo==4.5.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
scikit-learn==1.3.2
<<<<<<< HEAD
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
requests==2.31.0
pydantic-settings==2.4.0
=======
nltk==3.8.1
spacy==3.7.2
textblob==0.18.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
>>>>>>> Text-Analysis
